(feat) moved functionality from camera page to admin page
